refactor: LineController와 SectionController를 분리
