+++ README.md
@@ -204,7 +204,10 @@
   - 레벨 1을 하면서 equals, hashCode, toString을 언제 정의해야할 것인가를 주제로 많은 크루들과 이야기해본적이 있습니다.
   - 저는 equals와 hashCode가 있어도 코드가 길어지는 것 외에 단점이 없지만 없으면 HashMap, HashSet등에서 문제가 발생하기 때문에 있어서 나쁠것은 없다고 생각했습니다.
   - 함께 개발하는 다른 개발자가 해당 클래스를 사용하면서 당연하게 equals와 hashCode가 정의되어있다고 생각하고 개발한다면 문제가 발생할 수 있다고 생각하여 습관적으로 추가하는 습관이 생겼었습니다.
-- [ ] 라인 컨트롤러에서 섹션 서비스를 가지고 있는 이유가 무엇일까요?
+- [x] 라인 컨트롤러에서 섹션 서비스를 가지고 있는 이유가 무엇일까요?
+  - API에는 구간 추가, 구간 제거 기능을 하는 요청이 2개 있습니다. 이 요청의 URI는 "/lines/{id}/sections"로 시작합니다.
+  - 이 시작인 "/lines"가 이미 LineController 에서 사용되기 때문에 Section에 대한 요청도 여기서 처리할 수 있다고 생각했습니다.
+  - 하지만 노선과 구간이 다른 테이블에 저장되고 서비스로 다르게 사용하기 때문에 이 둘을 모두 LineController에서 가지면 너무 하는 일이 많아보이고 테스트도 너무 많아져서 둘을 구분했습니다!
 - [ ] 컨트롤러에서 섹션을 추가하면 반환되는 상태 코드가 200인 이유가 있을까요? + 삭제 시 200인 이유가 있을까요?
 - [ ] 이전 1, 2 단계에서는 Controller가 DAO를 알고 직접 변환하여 사용했는데 이 작업들이 서비스로 간 이유가 있을까
 - [ ] (schema.sql) 단순히 값으로만 선언해도 되었을텐데 외래키를 사용하신 이유가 있을까요?
