+++ src/main/java/wooteco/subway/ui/SectionController.java
@@ -1,0 +1,38 @@
+package wooteco.subway.ui;
+
+import javax.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import wooteco.subway.dto.LineResponse;
+import wooteco.subway.dto.SectionRequest;
+import wooteco.subway.service.SectionService;
+
+@RestController
+@RequestMapping(value = "/lines/{id}/sections")
+public class SectionController {
+
+    private final SectionService sectionService;
+
+    public SectionController(SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping()
+    public ResponseEntity<LineResponse> addSection(@RequestBody @Valid SectionRequest sectionRequest,
+                                                   @PathVariable Long id) {
+        sectionService.addSection(id, sectionRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping()
+    public ResponseEntity<LineResponse> deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
+        sectionService.deleteStation(id, stationId);
+        return ResponseEntity.ok().build();
+    }
+}
